+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: anoroita <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/22 17:01:00 by anoroita          #+#    #+#             */
-/*   Updated: 2018/05/30 12:26:49 by anoroita         ###   ########.fr       */
+/*   Updated: 2018/06/08 16:11:53 by anoroita         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int		i;
+	size_t	i;
 
 	i = ft_strlen((char *)s) + 1;
-	while (--i >= -1)
+	while ((int)--i >= -1)
 		if (s[i] == c)
 			return ((char *)&s[i]);
 	return (NULL);
